+++ CHANGELOG.md
@@ -5,13 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2025-09-19
+
+### Added
+
+- @capacitor/preferences を追加し、l10n.use() 使用時にストレージへ言語情報を保持
+- l10n に async getStoredLang() を追加
+
 ## [1.6.0] - 2025-09-12
 
 ### Added
 
 - npx tn-permissions を追加
 - IonPickerModule を追加
-
 
 ## [1.5.0] - 2025-09-12